Refactor: 酒の検索に関する処理をconcernsに移動
